+++ frontend/frontend/components/index.tsx
@@ -18,7 +18,7 @@ export { default as GetBalancesComponent } from './cycles/GetBalancesButton';
 
 // Overlay Components
 export { default as LoadingOverlay } from './overlays/LoadingOverlay/LoadingOverlay';
-export { default as ErrorNotification } from './overlays/ErrorNotification';
+export { default as ErrorNotification } from './overlays/ErrorNotification/ErrorNotification';
 export { default as DeleteConfirmation } from './overlays/DeleteConfirmation/DeleteConfirmation';
 
 // Copy Components
